WEEK02/Podcast Studio/requirements.txt
# Python dependencies for the Podcast Studio project
newspaper3k
openai
gradio
pydub
python-dotenv
